MainProject/requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.7.0
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
mysqlclient==2.2.7
pytz==2025.1
sqlparse==0.5.3
tzdata==2025.1
